+++ src/main/java/io/github/sakurawald/core/config/model/ConfigModel.java
@@ -85,6 +85,7 @@ public class ConfigModel {
         public CommandSpy command_spy = new CommandSpy();
         public CommandEvent command_event = new CommandEvent();
         public CommandDebug command_debug = new CommandDebug();
+        public CommandAdvice command_advice = new CommandAdvice();
         public World world = new World();
         public TeleportWarmup teleport_warmup = new TeleportWarmup();
         public TopChunks top_chunks = new TopChunks();
@@ -663,6 +664,10 @@ public class ConfigModel {
         }
 
         public static class CommandDebug {
+            public boolean enable = false;
+        }
+
+        public static class CommandAdvice {
             public boolean enable = false;
         }