+++ src/main/java/io/github/sakurawald/module/initializer/command_advice/CommandAdviceInitializer.java
@@ -1,0 +1,86 @@
+package io.github.sakurawald.module.initializer.command_advice;
+
+import io.github.sakurawald.core.auxiliary.LogUtil;
+import io.github.sakurawald.core.auxiliary.StringUtil;
+import io.github.sakurawald.core.auxiliary.minecraft.CommandHelper;
+import io.github.sakurawald.core.command.executor.CommandExecutor;
+import io.github.sakurawald.core.command.structure.ExtendedCommandSource;
+import io.github.sakurawald.core.config.handler.abst.BaseConfigurationHandler;
+import io.github.sakurawald.core.config.handler.impl.ObjectConfigurationHandler;
+import io.github.sakurawald.module.initializer.ModuleInitializer;
+import io.github.sakurawald.module.initializer.command_advice.model.CommandAdviceConfigModel;
+import io.github.sakurawald.module.initializer.command_advice.structure.CommandAdviceEntry;
+import io.github.sakurawald.module.initializer.command_advice.structure.CommandAdviceType;
+import net.minecraft.server.command.ServerCommandSource;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+public class CommandAdviceInitializer extends ModuleInitializer {
+
+    private static final BaseConfigurationHandler<CommandAdviceConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, CommandAdviceConfigModel.class);
+
+    public static final int KEEP_CLOSER_TO_PRIMARY_METHOD = 1000 - 750;
+
+    @SuppressWarnings({"ResultOfMethodCallIgnored", "unchecked"})
+    public static void processCommandAdvice(Object handler, ServerCommandSource source, String commandString, CommandAdviceType adviceType, CallbackInfo ci) {
+
+        // Log it.
+        LogUtil.debug("Process Command Advice: advice type = {}, command string = {}, command source = {}, handler = {}, ", adviceType, commandString, source.getName(), handler);
+
+        // Filter the advice entries by advice type.
+        List<CommandAdviceEntry> targetAdviceEntries = config.model()
+            .entries
+            .stream()
+            .filter(it -> it.adviceType.equals(adviceType)
+            || (it.adviceType.equals(CommandAdviceType.CANCEL_WITH_SUCCESS)
+                && adviceType.equals(CommandAdviceType.BEFORE_EXECUTING)))
+            .collect(Collectors.toCollection(ArrayList::new));
+
+        // Filter the advice entries by command source type.
+        targetAdviceEntries = targetAdviceEntries
+            .stream()
+            .filter(it -> !it.onlyValidWhenCommandIsExecutedByPlayer || source.isExecutedByPlayer())
+            .collect(Collectors.toCollection(ArrayList::new));
+
+
+        // Perform advice.
+        targetAdviceEntries
+            .stream()
+            .filter(it -> commandString.matches(it.matchCommandStringRegex))
+            .forEach(it -> {
+                // Cancel the executing of target command.
+                if (it.adviceType.equals(CommandAdviceType.CANCEL_WITH_SUCCESS)) {
+                    LogUtil.debug("Cancel the executing of target command {} with success for {}", commandString, it);
+                    if (ci instanceof CallbackInfoReturnable<?>) {
+                        ((CallbackInfoReturnable<Integer>) ci).setReturnValue(CommandHelper.Return.SUCCESS);
+                    } else {
+                        ci.cancel();
+                    }
+                }
+
+                // Replace captured-groups for commands.
+                Matcher matcher = Pattern
+                    .compile(it.matchCommandStringRegex)
+                    .matcher(commandString);
+                matcher.find();
+                List<String> commands = it.commands
+                    .stream()
+                    .map(cmd -> StringUtil.replaceGroupsPlaceholders(matcher, cmd))
+                    .collect(Collectors.toCollection(ArrayList::new));
+
+                // Execute commands
+                LogUtil.debug("Execute commands {} for {}", commands, it);
+                commands.forEach(cmd -> {
+                    CommandExecutor.execute(ExtendedCommandSource.asConsole(source), cmd);
+                });
+            });
+
+
+    }
+}