+++ src/main/java/io/github/sakurawald/module/initializer/command_advice/structure/CommandAdviceEntry.java
@@ -1,0 +1,16 @@
+package io.github.sakurawald.module.initializer.command_advice.structure;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+public class CommandAdviceEntry {
+
+    public String matchCommandStringRegex;
+    public boolean onlyValidWhenCommandIsExecutedByPlayer;
+    public CommandAdviceType adviceType;
+    public List<String> commands;
+}