+++ src/main/java/io/github/sakurawald/module/initializer/command_advice/structure/CommandAdviceType.java
@@ -1,0 +1,5 @@
+package io.github.sakurawald.module.initializer.command_advice.structure;
+
+public enum CommandAdviceType {
+    BEFORE_EXECUTING, AFTER_EXECUTING, CANCEL_WITH_SUCCESS
+}