+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (command_debug module) feature: add `/command-debug <command>` command, to see debug info of a command.
+- (command_advice module) feature: add `command_advice` module, to allow decorate existing commands.